fix(editor): block updates if YAML invalid
This allows the user to edit the YAML and protects
it from being overwritten when it is in an
intermediate state.

The dirty and valid scope values can be used to
signal when this control is not valid YAML and
when it is out of sync with the blueprint
topology.